An attempt at implementing wasi sdk's llvm "wasm assemble" codegen for retest.
This is super awkward, especially around setting up the import for linear memory. I think that's the proper thing to do here, that it's part of the externally-facing ABI for the wasm program. And really I'd cut corners with the wat-based codegen by not using an import there. So perhaps we ought to output the same definition for both, I'm not sure.