Remove dependence on internal.h in print/ir.h and lx/print/c.c.
print/ir.h only depended on it for FSM_SIGMA_COUNT in `struct
ir_state_table`, and IR_TABLE isn't actually implemented, so
this can be removed for now.

lx/print/c.c only needed internal.h because of print/ir.h, and
because of several direct accesses to fsm->statecount, which are
easily replaced by calls to fsm_countstates.