WIP: Extend reject hook to restore lx's old leaf callback behavior.
Changing the leaf and endleaf callbacks to accept and reject in #485
broke lx. This commit passes through the necessary information to
restore the old behavior. It works in the happy path, but needs further
testing.

libfsm's normal execution mode evaluates a DFA, character by character,
terminating either when the next character isn't a valid edge or the end
of input is reached (in which case it checks end state metadata). lx's
execution mode is a little different, because it's tokenizing -- instead
of reading to the end of input, it should consume as much consecutive
input that matches a particular token, then yield the token type and
suspend.

lx used to work by breaking abstraction and calling directly into
`fsm_print_cfrag` (overriding the leaf behavior to yield token types),
but when the callback interfaces shifted its internals no longer fit
what lx expected. Now the reject hook is passed the same state
metadata as the accept state, and the reject hook in lx checks whether
the end id is associated with a particular AST mapping and token type.

Further things to check:

- There's a special case for the end of input, because it can't unget
  the next character. Ensure a token at EOI is tagged correctly.

- It currently doesn't check if the state is an end state, just whether
  it has at least one endid. This shouldn't matter, but it should check.

- It also doesn't handle multiple end IDs. lx seems to report errors for
  inputs that can match multiple tokens, so this may be unreachable.

- It's only implemented for the "c" output format, but would probably
  be usabled by others without a lot more work. In particular, kate
  mentioned it'd be good to be able to use vmc output for lx.