Optionally save linkage_info during NFA construction in re_comp.
This adds a new flag for re_comp, RE_SAVE_LINKAGE_INFO. If that's set,
then save information about the unanchored start and end loops, the
end loop end, anchored starts, and anchored ends while building the
NFA from the regex. Previously, this info was recovered by analyzing the
NFA structure inside of fsm_union_repeated_pattern_group, but it seems
less brittle to save those details during construction.

This commit adds the flag, saves the info during construction, and
after doing analysis it checks the results match. Since there is one
case where it doesn't (eager_output_fr3.c), print a message and override
it. This was actually a bug: that particular edge case leads to
detecting the same state as both the unanchored start and end loops. In
that specific case, it seems to be harmless, but this commit also adds
an assert that the USL and UEL are not the same state.

From now on, fsm_union_repeated_pattern_group depends on all of the NFAs
passed to it having the linkage_info struct populated. They should come
fresh from re_comp, without determinising or other transformations. I
added an assert at the end of determinisation checking that the DFA does
not have a linkage_info struct.

Note: fsm_clone currently doesn't copy the linkage_info. Maybe it should.