Copy some fields from linkage_info, remove analysis.
They need to be copied over (transferring ownership of the state sets)
because the linkage_info struct gets freed during the call to fsm_merge.

This removes some code that tended to reconstruct the same information
by analyzing the NFA structure, but was significantly more compliacted,
and in at least one case it led to the wrong result (albeit harmlessly).
It's simpler to just save the relevant state IDs during NFA construction.